fixed rakefile blowing up on controllers
I think it was due to some state hanging around in the compiler causing compilation
to fail either when the form_helper is compiled in the same session as the contacts cntrler
or when the contacts controller is compiled multiple times

I fixed it by calling mirahc from the commandline on controllers. This makes compiling much slower,
but it also means that it will work.